storemodel: requires `KeyStore` to cheaply clone itself
Summary:
See previous changes (D64561396) for context. Basically, we'd like to avoid
`iter` APIs to return `BoxIterator` that has a lifetime relation with the store
itself.

Note: We cannot simply write `trait KeyStore: Clone` because the return type of
`clone()` is a concrete type, not a trait object. Rust will then disallow
`KeyStore` to be used as a `dyn` trait object. Trait object `dyn KeyStore` is
widely used for abstraction.

Reviewed By: muirdm

Differential Revision: D64561516

fbshipit-source-id: 22126114e045185dbad660fd0ca8b4f3aa361811