eagerepo: land_stack supports multiple commits
Summary:
This diff uses pushrebase_one() to support pushrebasing a diff stack with multiple commits. Currently, it does not support concurrent land_stack calls. We will handle it in a later diff.

The wireproto does not work for this test case, it seems didn't do the rebase first. I will fix it in a separate diff.

Reviewed By: quark-zju

Differential Revision: D64540084

fbshipit-source-id: 915bbce72da23a3075d6c87cb4afd124e590c9fd