Fix the API for per-repo vs global queue
Summary:
## This stack

S460221 followups to fix bugs around sharding as well as general brittleness.

## This diff

Back in the days before my changes, we always assumed requests were associated with a repo.

I added support for a "global queue", i.e. requests not associated with any repo, which are identified by `repo_id=NULL`. However, I got the query API wrong. I meant to support either of these:

* `repo_id IN (...)`
* `repo_id IN (...) OR NULL`

But instead I went with

* `repo_id IN (...)`
* no condition on repo_id

This was a big mistake, as it made it too easy to misconfigure.

This diff changes the meaning of the `repos` argument: if `None` it means the current queue object *only* supports global requests.

Reviewed By: RajivTS

Differential Revision: D64692009

fbshipit-source-id: 2cb4d03b8d0c1f26b18c5b1007728c7b1a665fd9