Start an executor for the global queue, even when sharded
Summary:
## This stack

S460221 followups to fix bugs around sharding as well as general brittleness.

## This diff

Earlier in the stack we broke out processing of the global queue to a separate executor, but we only did so for the unsharded case, as that's all we have in integration tests. Now do it for the sharded case.

Reviewed By: RajivTS

Differential Revision: D64695135

fbshipit-source-id: 4a3aaefe894931bffdcf03add07403461d8708de