Make changed files list more compact
Summary:
The list of files often takes up a lot of space. We can make this slightly more compact to increase how many files you can see without scrolling.

We seem to have set an explicit line-height here. Reducing this seems to work in all the places we show this file list. This value seems to be a good middle ground which does not cause buttons to overlap and gives a small amount of spacing between files, but still looks reasonably good. I'm trying to avoid being too compact here.

Is it too much? Should we take a less aggressive value instead?

Reviewed By: quark-zju

Differential Revision: D64706555

fbshipit-source-id: c8ff106d27430bad268a4e2465541e0772589333