Handle redirection symlink in windows: Delete redirection target mapping when running `redirect del`
Summary:
# Background
We have seen tools like Unity messing up with already existing redirection symlink by deleting and/or creating a folder/file instead of symlink.

# Solution
When a create/delete operation is triggered for a redirected path, handle it in a way such that redirection stay intact.

# This change
Read redirection target mapping saved in checkout config as part of D63945643 and load it to CheckoutConfig.cpp so that it can be used to handle create and delete projected-fs notification on redirected path.
This diff is to delete the redirection mapping before actually deleting the symlink, so that normal deletion can progress.

Differential Revision: D64344923

fbshipit-source-id: 0f84710ea2049fc12b4efb0a7c8591e8047ec6b9