patch: highlight the trailing white space in diff view
Summary:
This diffs make the worddiff highlights the trailing white space. Trailing white space is hard to see, highlighting it will make it easier to review the changes.

Another alternative solution is to change the "tokenizer" to make it cannot have two consecutive tokens like `(True, b" "), (False, b"\n")`. This approach is likely faster but less robust than the change proposed in this diff.

Reviewed By: sggutier

Differential Revision: D64603891

fbshipit-source-id: 608939288837ebae4bb522915e44a3b2df6480d5