patch: refactor names in diffsinglehunkinline()
Summary:
This diff renames `nextisnewline` to `isprevnewline`. `nextisnewline` actually
means the current token will be on a newline. Let's rename it to
`isprevnewline`, which means if the previous token ending with a newline.

The motivation of this change is that I will add a `isnextnewline` variable to
check if the next token is a newline. This will be used for handling trailing
white space issue.

Reviewed By: quark-zju

Differential Revision: D64603889

fbshipit-source-id: d7a2dfe6f2b54e31b1201d033ee50e6572e05ade