PathComponent: implement convenience traits
Summary:
# This diff

Some convenience traits were missing for PathComponentBuf, but present for RepoPathBuf and RepoPath. This diff adds some of those trait impls for PathComponentBuf to:

1) reduce the gap between the two types (hopefully encouraging people to use the correct type for their usecase in the future)
2) Make the change I'm introducing in the next diff much easier to implement

# Context

Several places in the code use RepoPathBuf in places that PathComponentBuf should be used. I'm going to fix a couple of those instances to help avoid future confusion

Reviewed By: quark-zju

Differential Revision: D64565160

fbshipit-source-id: fe7a0c38558a54cf421265d65abf965956e28535