AugmentedTrees: use PathComponentBuf instead of RepoPathBuf
Summary:
# This diff

Migrates AugmentedTrees to use PathComponentBuf instead of RepoPathBuf. The latter is used to describe paths that are relative to the repo root. In the case of AugmentedTrees, the path is supposed to be the basename of the entry (i.e. relative to the entry's parent). Therefore it doesn't make sense to use RepoPathBuf, and it simply adds confusion as to what value the field holds

# Context

Several places in the code use RepoPathBuf in places that PathComponentBuf should be used. I'm going to fix a couple of those instances to help avoid future confusion

Reviewed By: quark-zju

Differential Revision: D64565163

fbshipit-source-id: 56e500b981654ed9d2d70656409977c9a18caa4e