mercurial: use Bytes instead of Vec<u8> for mercurial changeset structs
Summary:
Currently the Mononoke Mercurial types use `Vec<u8>` to store byte strings.  These have the unfortunate side-effect of debug-printing as a list of integer values.

The `bytes::Bytes` type has a good `Debug` impl that prints as a Rust byte string literal.  It's also probably a better type to use.  Switch to that for the Mercurial types.

Reviewed By: lmvasquezg

Differential Revision: D64703854

fbshipit-source-id: 46d8a32aeb8332610f15f6ae41b7de7e8f912439