cas: reorder cas fetches to happen before local cache fetches
Summary:
In CAS mode we don't want to waste time querying the local indexedlogs if we are going to hit CAS. Improve by re-ordering the CAS fetch to happen before the on-disk cache check.

Note that I left the local in-repo read before CAS. This is so when data is available only in the in-repo local store (i.e. you just committed it), we don't perform a CAS network request unnecessarily.

Differential Revision: D63799142

fbshipit-source-id: d39e42786b30ee37d02f91d07b201c94acb779dc