debugnetworkdoctor: change to an optional-repo command
Summary:
Load the local repo's config, if available. The local repo's config could change the server hostname or other important inputs into the network doctor.

I noticed this when trying to use the network doctor to diagnose issues connecting to a Mononoke in AWS. It was using my system defaults instead of the per-repo configuration.

Reviewed By: quark-zju

Differential Revision: D64699808

fbshipit-source-id: 3c89869a903d6fc738b5f51b8825aff279b07262