networkdoctor: handle http proxies better
Summary:
The current TCP connectivity check is invalid with an http proxy in play since we try opening a TCP connection directly to the server (which is expected to fail if you must go through an http proxy to reach the server).

Fix by proceeding to the http specific connetivity test even if the lower level tcp check failed. If we got a more specific http related error (like bad response), use that error instead. This indicates we were able to connect to the server over http, so we should ignore the previous direct tcp connection error.

Reviewed By: quark-zju

Differential Revision: D64699811

fbshipit-source-id: 1668f07f724b5463eb3550687a3baf6c10b65c9d