cas: fix fetching hg file content
Summary:
Hg file content includes the raw file content along with any hg content header (e.g. file copy metadata).

Normally the file store contains the hg file content so it is all one piece. However, CAS only contains the pure content, so we need to fetch the hg content header separately. The header is stored in the file AUX data, but a couple things were preventing it from being propagated:

1. When requesting hg content (but not aux data), the aux data (containing the hg content header) was getting "masked" off the result because the request apparently does not want the aux data. Fix by making the `mask` method know that the content header might come from either the content or aux data, and retain them accordingly.

I was tempted to simply get rid of the masking, but I think its purpose is to drop the file content from the result if only aux data was requested (to save memory). The file content would be present when requesting aux data if it was used to "derive" the aux data.

2. The `hg_content()` method was implemented on LazyFile, but for CAS LazyFile only contains the pure content. Fix by moving `hg_content()` up one level to `StoreFile`, which contains both the content and aux data. For CAS, the new `hg_content()` will combine the content and header together.

Reviewed By: liubov-dmitrieva

Differential Revision: D64717408

fbshipit-source-id: 7b2db16ee60d621b40f2ecbbf208d438ae514976