derived_data_service: simplify error handling in build_underived_batched_graph
Summary: Right now when trying to enqueue a derivation item into the queue and failing, in some cases we check that the item was derived and in other cases we don't depending on the type of the error. This diff makes it so that all errors other than `ItemExists` are handled in the same way by first checking if the item was derived and retrying to enqueue if it's not. The less cases we have to reason about the easier it's to debug corner cases.

Reviewed By: RajivTS, singhsrb

Differential Revision: D64833459

fbshipit-source-id: 23cfa17b797f50519abcc89d553d8ee15e17998f