Handle redirection symlink in windows: Handle proejcted-fs pre-delete dir notification
Summary:
# Background
We have seen tools like Unity messing up with already existing redirection symlink by deleting and/or creating a folder/file instead of symlink.

# Solution
When a create/delete operation is triggered for a redirected path, handle it in a way such that redirection stay intact.

# This change
Handle proejcted-fs pre-delete dir notification: When a pre-delete notification is received, path is redirected and is a symlink. Do not delete it, instead best try deleting content of target directory and return error so that file system can reject the delete operation.
We will handle the case of redirected path being file or directory in post delete notification and seaprate diff.

Reviewed By: genevievehelsel

Differential Revision: D64366283

fbshipit-source-id: 81b4f6eb6c5e429065d43cccf37f514793618a29