Handle redirection symlink in windows: Read latest redirection target mapping config in daemon
Summary:
# Background
We have seen tools like Unity messing up with already existing redirection symlink by deleting and/or creating a folder/file instead of symlink.

# Solution
When a create/delete operation is triggered for a redirected path, handle it in a way such that redirection stay intact.

# This change
Read redirection target mapping saved in checkout config as part of D63945643 and load it to CheckoutConfig.cpp so that it can be used to handle create and delete projected-fs notification on redirected path.
D64151373 was made to read the redirection config but since these redirection configuration can change anytime and projected fs notification needs to know the current state. This change it to read latest redirection target mapping. Given file size is small for this config and OS will put it in RAM for frequent reads, I dont see any major perf concern. Also, I dont expect a lot of delete operations on symlink.

# alternatives considered
1. We could keep a last_updated file time and refresh the redirection-target mapping only if it changed. Problem with this approach is that the mapping will get overridden and the iterator already using it will behave mysteriously.
2. Take a write lock on mapping. This will slow down projected-fs notification execution -> slowed files operations and possible deadlock?
3. hybrid approach: use last updated time for write and use read semaphore- little complicated. But can send a separate diff with this approach.

4. (Happy to know if there could be other easier alternatives.)

Reviewed By: genevievehelsel

Differential Revision: D64357995

fbshipit-source-id: dca5aaa13498b2cd71cfe7a2885e483debcd0279