hook_manager: don't use common scuba builder to generate hooks
Summary:
The generation of hook execution futures for hooks currently shares a single scuba sample builder (via `&mut`).  This means we accidentally pollute the builder's values as we construct the list, and all hooks get run under the name of the last one.

Switch to cloning and filling in the hook name just before we create the future.  Make it clear we don't expected to mutate the main builder by switching to immutable references.

The `ScubaSampleBuilder` API is pretty annoying, but fixing that will be for another day.

Reviewed By: gustavoavena, lmvasquezg

Differential Revision: D64901151

fbshipit-source-id: 7832684769c4b068658c8e1f4e5d4c76d251f971