merge: avoid unnecessary condition checks
Summary:
This diff avoid unnecessary condition checks by combining multiple `if`
statements to a chain of `if-elif`. We can do this because the variable
`m` does not change and can only match one of the conditions.

With this change, if it matches the first `if` condition, it will not check
the following conditions.

Reviewed By: muirdm

Differential Revision: D64901699

fbshipit-source-id: 0e88a526264dd3b3de3938d513dae39267381f6e