merge: use ACTION_X constants instead of strings
Summary:
This use the ACTION_X constants in more places and also defined some new action types.
The motivation is that I found it's hard to understand the hard-coded string when reading
merge related code and the string literals are not `grep` friendly.

The naming of the constants is based on the Mercurial upstream change by Gregory Szorc: https://phab.mercurial-scm.org/D2701

Original message:

We finish up establishing named constants in this file with
actions.

I remember scratching my head trying to figure out what this
code was doing as part of addressing a recent security issue with
subrepos. Having the named constants in place definitely makes
things easier to read.

I'm not convinced the new constants have the best names (I'm not
an expert in this code). But they can be changed easily enough.

Also, since these constants are internal only, we might want
to change their values to something more human readable to
facilitate debugging. Or maybe we could employ an enum type
some day...

Reviewed By: muirdm

Differential Revision: D64877735

fbshipit-source-id: 0bb0aa5610cb8d372e730e5b8d1cc49d93879a2c