derived_data_service: stop searching for derived ancestors past batch parents in build_underived_batched_graph
Summary: The logic in D64899662 for finding the underived part of a batch is not completely sound because the parent batches could be not derived yet as the code only guarantees that they have been enqueued in the derivation queue. This diff fixes the logic by excluding changesets with generation number smaller than the current batch's root.

Reviewed By: singhsrb

Differential Revision: D64959928

fbshipit-source-id: 56fb5357a12841bd6b2f02eb8bfa0dab682d34cd