derived_data: remove visited HashSet from DerivationDependencies::check_dependencies
Summary: The `visited` `HashSet` in `DerivationDependencies::check_dependencies` is only used to deduplicate entries in the list of types defined using the `dependencies` macro. This means that for it to have any value someone must accidentally write the same type twice i.e. write `dependencies![X, X]`. We don't currently have any instances of this and it's meaningless to do so let's remove this unnecessary deduplication logic.

Reviewed By: lmvasquezg

Differential Revision: D64962364

fbshipit-source-id: db3645b9e9bc2c79641b83b043a624566a7c6c52