scmutil: print UncategorizedNativeError as Abort error
Summary:
Oftentimes, the Rust land uses `bail!` or `anyhow!` that crashes the program
with `UncategorizedNativeError`, which is a suboptimal UX. Just treat it like
an `Abort` error.

Ideally, we can test if the error type is from `bail!`. However, it is a
private type with a type parameter:

  https://github.com/dtolnay/anyhow/blob/6c52daaa79eb22279248d80bebedc6c14bbb84ec/src/wrapper.rs#L11

So it's hard to figure out exactly.

Reviewed By: sggutier

Differential Revision: D64929157

fbshipit-source-id: 852365a96fb6608766f95659d06f497035399a0c