FBID is not unixname
Summary:
In some cases (pushes via SCS initiated by humans) we are being provided two
`USER` identities: one of which is unixname and other fbid. If fbid is the
first one this function will return it without looking for the other one.

This unixname is used by verify_integrity hook and compared with the commit
author (so people don't push commits that are not theirs) so FBID is natually a
mismatch there.

Reviewed By: gustavoavena, singhsrb

Differential Revision: D64973239

fbshipit-source-id: c6cba928a45fa779f559ee3356d45973eaef0db4