rust-remove: check active eden mount
Summary:
Check if the given path is an eden repo root.

## This Diff
We implement another state of the state machine: **ActiveEdenMount**.

Under Linux/MacOS, an active eden mount will include a ".eden" folder which contains a symlink pointing back to the repo root.

This is used as the criteria inside **Determination** State.

Reviewed By: MichaelCuevas

Differential Revision: D64739286

fbshipit-source-id: 0834998a6ee5364a81db5498cf69de345936d1a7