changelog: switch to Rust commit fields parsing
Summary:
This makes `changerevision` support the `Git` format. A config option is added
in case we need the old behavior.

Lots of places use the commit extras for "committer" and "author_date".
It's a bit complicated to migrate all users to dedicated fields on
`changelogrevision`, `context` (including constructors of various `context`
types). So I just made it convert the Rust "committer" back to "extra" for now.

The "date" field must be one of "author_date" or "committer_date" now. It
cannot be a 3rd different date. Therefore some tests are changed.

Reviewed By: muirdm

Differential Revision: D64833696

fbshipit-source-id: 45dbaa1da00be964a35c844db2487fa7d19c0cae