merge: abort merge when there is a conflict with subtree copy
Summary:
This is mainly for subtree copying to an exiting directory case. For example:

* B created a subtree copy from `foo` to `foo-stable` based on an ancestor commit.
* C (from another user) added a new file to `foo-stable` and landed the commit first
* when rebasing B onto C, there is a conflict for the `foo-stable` directory. We don't
  want to merge `foo-stable/new_file` into B automatically, this might conflict with B's
  intended state.

```
C add foo-stable/new_file (remote/master)
|
|  B subtree copy foo to foo-stable
| /
A
:
```

This should be an uncommon use case, so as the first step, we will just abort the rebase
for the commit, and ask users to re-create the the subtree copy by checking out to a
commit after C. If users need the change of C, users can do a `subtree graft` or
`subtree merge` to merge it to the `foo-stable` directory. Later, we will handle this
process automatically.

Reviewed By: quark-zju

Differential Revision: D64901698

fbshipit-source-id: 65d1f310196b27c78959e2adbf5d1253f0e346a5