Handle redirection symlink in windows: Delete redirection target mapping when running `eden unmount`
Summary:
# Background
We have seen tools like Unity messing up with already existing redirection symlink by deleting and/or creating a folder/file instead of symlink.

# Solution
When a create/delete operation is triggered for a redirected path, handle it in a way such that redirection stay intact.

# This change
Read redirection target mapping saved in checkout config as part of D63945643 and load it to CheckoutConfig.cpp so that it can be used to handle create and delete projected-fs notification on redirected path.
This diff is to clear the redirection mapping before deleting symlinks as part of `eden unmount`, so that the deletion can succeed.

Differential Revision: D64882112

fbshipit-source-id: 10613b833bb8e3397b34b31e9c3d9d4bc7ef252b