cas: fix metric accounting for duplicate keys
Summary: If multiple `(path, node)` keys map to the same CAS digest, we were recording the number of digests as "fetch" (i.e. attempt) count, but the number of keys as the "hit" count. This could lead to hit count greater than fetch count. Fix by recording the number of keys as "fetch" count.

Reviewed By: liubov-dmitrieva

Differential Revision: D65067201

fbshipit-source-id: 7db173e8de7d41787f78beb5e9342d90e6331adc