edenapi/types: accept git SHA1 hashes
Summary:
Update the hash verification to also check Git hashes. Ideally, we can figure
out which format (git or hg) in related code paths. But that seems to require
too many changes (ex. if we want to track the format in the Tree or File
struct, we might have to update the wire types and potentially server-side
logic).

Reviewed By: zzl0

Differential Revision: D64948071

fbshipit-source-id: 5b93d45e1f33d770d52545792178876d152b2c7b