cas: fix file type when converting cas tree to hg
Summary:
When CAS is enabled and you fetch a manifest tree through the manifest-tree crate (used by Sapling, but not EdenFS), the bytes go on this journey:
1. Fetch augmented tree bytes from CAS
2. Deserialize into AugmentedTree structure
3. Serialize AugmentedTree into hg tree bytes
4. Deserialize hg tree bytes (in manifest-tree crate)

I previously had fixed non-regular file serialization for 1. and 2., but hadn't noticed 3 was encoding all files as regular. Fix it to write the "x" flag for executables, or "l" for symlinks.

Reviewed By: quark-zju

Differential Revision: D65166996

fbshipit-source-id: fd875b8140f0ed25851920b00644a8f9fd9ffc4c