cas_client: return false instead of error in is_augmented_tree_uploaded for missing augmented trees
Summary: Whenever we call `is_augmented_tree_uploaded` it's possible that the hg augmented tree hasn't been derived yet which will result in returning an error instead of returning false. This is unintended as it's expected in some cases for the augmented trees to not be derived, like in the cas uploader if it's racing with warm bookmarks cache.

Reviewed By: singhsrb

Differential Revision: D65204099

fbshipit-source-id: 9e041f9320e67241204edee9efb016f50f14a018