clone: fix repo-specific dynamic config for Python clone
Summary:
I noticed the Python clone wasn't getting repo specific dynamic configs (based on the repo name of the repo being cloned).

Fix by forcing an in-line recomputation of the dynamic config if the in-hand repo name mismatches with the repo name in the header of the on-disk file.

This is an edge case that occurs only because the Python intializes the dest repo before writing out the hgrc. This causes the repo to load (and cache) config without knowing the repo name.

Reviewed By: zzl0

Differential Revision: D64498409

fbshipit-source-id: 56228dd8e1963b60c00e2663f6b3a5fc738ba010