context: make changectx.changeset() to return an object instead of a tuple
Summary:
We are planning to use the `changectx._changeset.files` in the D65217338 when
uploading changesets via SLAPI (EdenAPI). But we don't want to use the private field
`_changeset`. So I am planning to use `changectx.changeset()` function, however it
returns a tuple, I have to use `changectx.changeset()[3]` to get the `changectx._changeset.files`,
which is not readable.

This diff make `changectx.changeset()` return an object, then we can access the `files`
field with `changectx.changeset().files` instead of `changectx.changeset()[3]`.

There is no usage of `changectx.changeset()` by searching `.changeset` in the code base.
It should be safe to change it.

Reviewed By: muirdm

Differential Revision: D65219905

fbshipit-source-id: 6e3dc34ac88bc0a2447b49123dc3a6d2e36a85a5