hgtime: remove more deprecated functions
Summary:
`i64::min_value()` and `i32::max_value()` have been replaced by `i64::MIN` and `i32::MAX` constants. We can therefore change the code to refect that.

However, upon changing the `i64::min_value()` call, I realized the comparison that was happening was unecessary. This is because `time.timestamp()` is an i64, and therefore it will ALWAYS be >= `i64::MIN`. We can remove this unecessary comparison as well and simplify the code a bit.

Reviewed By: quark-zju

Differential Revision: D65224042

fbshipit-source-id: a3fa02a780456c5992b6b157b04e0c3c9282deb4