clone: kill legacy CLI flags and codepath
Summary:
Kill the legacy `--pull`, `--rev` and `--stream` CLI flags, and remove support for `--no-shallow`. None of them are used anymore. Note that `--stream` is still implicitly enabled in certain cases, but there doesn't need to be a CLI flag.

I removed the fallback `exchange.pull()` codepath from the Python clone. This is definitely not used in production, and I think tests shouldn't need it anymore either.

This change is aggressive but should be pretty safe. We don't use the Python clone at all in production.

Reviewed By: sggutier

Differential Revision: D65102302

fbshipit-source-id: 4ddbe75f9f1f80fd2944ff4d3958df93843acf4c