clone: kill uses of --stream
Summary:
Kill uses of streaming clone in tests. This is a legacy clone mode we don't use anymore.

I changed the test for streaming clone itself to use the Rust clone in non-commit-graph mode. This performs a streaming clone of the changelog (which is the only remaining use we have in production).

Reviewed By: sggutier

Differential Revision: D65102304

fbshipit-source-id: 4210c565754642e53baa88a62487fb6b41b1dd97