Don't code split css for now
Summary:
Problem: production builds of the vscode extension internally had missing CSS.

The reason is that our build now includes both the ISL webview and the inline comment webview. By building in the same vite instance, it does code splitting.  This is good for code reuse.

Our vscode extension has its own copy of the html that it customizes with CSPs and other values. This html is separate from the `addons/vscode/webview.html` and `inlineCommentWebview.html` files that are the actual sources of builds in vite. We were relying on the output html to be very similar to the input, and then manually patching in the right file names according to the build files.

This began breaking down once we had two entry points. Once the inline comment webview starting using more ISL CSS, the shared CSS was moved to a new shared chunk. However, our vscode html would only try to load the single css chunk corresponding to its entry point. This meant that ISL was missing a lot of styles.

The fix in this diff is to turn off code splitting for css in the rollup options. This means we always output exactly one css file: `dist/webview/res/style.css`. This way, each webview can specify that as its CSS and it's guaranteed to be correct regardless of other webviews.

The downside of this approach is that both webviews get ALL the css. This is not ideal for the inline comment webview, which needs to be as small as possible for improved load times. This is quite a serious downside, and we need a new approach to solve this for real.

This will be done separately, but roughly I see a few options:
- try harder to get vite to code split in the exact way we want, where there's minimal chunks, or chunks use import syntax so the HTML only needs a single css `<link>`.
- change how `htmlForWebview` works such that it reads the output `dist/webview.html` file generated by vite. This file includes all the correct CSS files linked for us. The issue here is that we'll need to make the HTML file generated by vite have all the proper stuff needed for vscode, namely a proper CSP / nonce, and various placeholder HTML, additional CSS, additional scripts, and more. This is probably the best way forward. An additional benefit here is that we get support for dev mode without needing a majorly different code path / HTML implementation.
- similarly, we could read the build `manifest.js on` and try to add in all the appropriate css files to the HTML in htmlForWebview. This is probably more fragile.

Differential Revision: D65310448

fbshipit-source-id: 8298b68623d9a2a2aec80724ebc72b3caf11175a