derived_data_service: deduplicate dependencies before enqueuing in the derivation queue
Summary:
This is for the case where we have a merge commit that has one merge parent which is itself a parent of the second merge parent:
```
A
|\
| \
B |
| /
|/
C
```
The dependencies of [A] could be [B, C] and [C] which both have the same derivation queue ID because we use the root of batches. This can result in all enqueue attempts to fail because we will try to create the same dependency zelos node twice. Let's deduplicate dependencies before attempting to enqueue to prevent this.

Differential Revision: D65334032

fbshipit-source-id: 5d3ef90983cd2bfac012056b76ba897455baf066