ejeca: add a skeleton for the library
Summary:
In order to show progress all the way from EdenFS into ISL we need some a   communication channel between Sapling and ISL. NodeIPC is a great candidate for this, and newer versions of the execa package support this.

Unfortunately, upgrading to newer versions of execa is really difficult since:
- They introduce a few minor breaking changes to its API (e.g., timeouts for killing commands)
- Newer versions of execa are [ESM only](https://github.com/sindresorhus/execa/issues/489), and making Jest work with this is really annoying.
- The [CJS fork of Execa](https://www.npmjs.com/package/esm2cjs/execa) hasn't been updated in a while and doesn't have the IPC capabilities we want

We still want the important bits of execa, however. Changing execa for direct calls to the native `child_process.spawn` is not something we want to do because that would be tricky and we would have to be really careful when mocking commands through Jest.

Because all of that, we've decided on creating a minimal clone of execa which only has the capabilities we already use in our codebase. This diff adds the skeleton for a new library called "ejeca" which does precisely that.

Reviewed By: evangrayk

Differential Revision: D65138584

fbshipit-source-id: 6591a31f8d5163e3285a5b1ef3e9e375faafd056