Add logging for exceptions happening in InvalidateChannelEntryCache
Summary:
# Context
On windows, the checkout failure/corruption rate is 5% higher than on other platforms. One of the contributors to this file locking during checkouts. There was a previous investigation linked below. However, we do not currently have data on what

# This Diff
This diff adds logging to the exceptions that occur in InvalidateChannelEntryCache. This is intended to track the occurence rate and location of file conflict errors, but will also log other errors that may happen.

# Technical Details
Add new log event. This log tracks the inode_logpath, which is the relative path to the file within the repo, and the exception message.
Log when InvalidateChannelEntryCache sees an exception.

# Discussion Points
1. Do we need to add collumn limits on the scuba table

Reviewed By: jdelliot

Differential Revision: D65072053

fbshipit-source-id: 7abef4b347ce001cb818ffa2e08d2a84d28c6538