work around flakiness in submodule expansion test
Summary:
The submodule expansion tests can be flaky because the bookmark update for the submodule expansion might not be immediate, so the clone might end up with a different commit for the master bookmark.

Workaround this by adding a small sleep for now.  We should improve the test so that this isn't an issue in the long-term.

Reviewed By: liubov-dmitrieva

Differential Revision: D65333796

fbshipit-source-id: df0fba562fd0720ce79d9a14b5cfc0f79572eac0