rust-remove: remove client config dir
Summary:
Implemented InActiveEdenMount state for the rust remove state machine.

In this state we first remove the client directory for the mount point. Then we remove the mount point entry from the config JSON file.

Note: before writing to the JSON file, we need to obtain the file lock first. This is to make sure we are the only process that is accessing that file.

Reviewed By: jdelliot

Differential Revision: D64945986

fbshipit-source-id: 37f411a55d6b5eae58369fcec7d5a05a703f2311