Fail open on JK read failures
Summary:
After D65261616 was pushed, it broke backsyncing completely, because when applying changes to it, I changed the JK name by mistake.

This led to backsyncing failing completely.

Let's not crash everything on JK read failure. Instead, log something so we can debug unexpected behaviour in the TW logs, but continue with the default value of false.

Reviewed By: lmvasquezg

Differential Revision: D65414558

fbshipit-source-id: 14c24f7201466bdcab7f7f1d37b823111274ccf0