blobrepo_hg: remove create_bonsai_changeset_hook
Summary: The `create_bonsai_changeset_hook` mechanism is only ever used to verify against an origin repo, and this is defined in the same place as it is used.  The hook mechanism is unnecessarily complicated and makes evolving the `CreateChangeset` interface harder.  Remove it in favour of clear-coding origin verification.

Reviewed By: clara-9

Differential Revision: D65059468

fbshipit-source-id: be60d0fce6c16ebabd51309e1cef9417170727c9