gitimport: On re-import, delete Mononoke bookmarks that are no longer in git
Summary:
This is something I'm hitting in prod when importing many repos and the import lasts a while:

If anyone deletes a branch upstream in Metagit, `gitimport` ignores it in the re-import and I need to follow-up manually cleaning-up before things can be in sync again.

This diff automates it.

Reviewed By: YousefSalama

Differential Revision: D65426363

fbshipit-source-id: 0470a121a8c937fcf971441e58197f2108fc5446