gitimport: test: Reproduce an issue with re-import after ref deletion
Summary:
This is something I'm hitting in prod when importing many repos and the import lasts a while:

If anyone deletes a branch upstream in Metagit, `gitimport` ignores it in the re-import and I need to follow-up manually cleaning-up before things can be in sync again.

Let's automate it. The first step is to repro in an integration tests, which is what this diff is about.

Reviewed By: YousefSalama

Differential Revision: D65426364

fbshipit-source-id: 8bd1310dcfbf8f8f23804751dac89136cffcac86