mononoke_types: add definitions of content manifest types
Summary:
Introduce `ContentManifest`. This is the replacement for fsnodes that is sharded, and simplified to exclude additional roll-up data like counts and hashes.  This will make individual manifest blobs smaller so that they are always cacheable.  Additional data like counts and hashes will be added in secondary manifest types later on, so that we can completely replace fsnodes.

This diff adds the mononoke types and conversions to/from thrift.

Reviewed By: andreacampi

Differential Revision: D55075373

fbshipit-source-id: a9ab52e577cc1d19d0a1ee7621ee329f9243cad7