content_manifests: add thrift definitions
Summary:
Introduce `ContentManifest`. This is the replacement for fsnodes that is sharded, and simplified to exclude additional roll-up data like counts and hashes.  This will make individual manifest blobs smaller so that they are always cacheable.  Additional data like counts and hashes will be added in secondary manifest types later on, so that we can completely replace fsnodes.

This diff adds the serialized thrift types.

Reviewed By: clara-9

Differential Revision: D55063071

fbshipit-source-id: 2cae6271aba102abb78f6b9c6061638411b51713