scmstore: expose hg content header
Summary:
Expose the hg content header data via LazyFile::file_content() method. This makes it easier to get the hg content header when it is attached with the "pure" content.

I'm going to use this to allow FileStore::get_rename_iter to avoid fetching the file content (it only wants the hg header).

Reviewed By: quark-zju

Differential Revision: D65376833

fbshipit-source-id: bc5d052256769735c57abab3a956c583a0f84f7c