remotefilelog: stop prefetching parent commit files
Summary:
The remotefilelog extension overrode the "commitctx()" method to prefetch the parent commit's file contents for files added/modified in the new commit. I don't think in general we use the parent commit's file contents when committing since we do cheaper filenode comparisons.

Note that for the O(1) "subtree copy" operation where we reuse tree nodes at different paths, we do expect to fall back to content comparison in `remotefilelog.cmp()`. If that becomes an issue, we should address with a specific prefetch for the subtree copy case.

Reviewed By: quark-zju

Differential Revision: D64947721

fbshipit-source-id: 2713175ef883f0d9ea668cc19472859e4d9f1c1d