commit: reuse filenode in more cases
Summary:
Before adding a file to the filelog, we check if we can reuse an existing entry. Previously that logic was gated by a `newfilenode in flog.nodemap` check, but that isn't optimal beacuse the nodemap only checks the local filelog (i.e. does not do a remote query), so it misses opportunities to reuse the filelog entry and save some work.

I realized that the nodemap check probably isn't required at all. We already check the new file node's parents, so if we assume that being able to look up a node's parents means the node itself exists (which I think is true), we don't need to directly check for node existence.

Reviewed By: quark-zju

Differential Revision: D65376845

fbshipit-source-id: 1cea3f1c8a9708b1bd861d63786c4abc8209c8c5