merge: avoid repo.commitpending() for in-memory merge
Summary:
For in-memory merge we don't do a real "update" and certainly don't use the Rust workers (which can only write to the real filesystem), so we don't need to spend time flushing pending data.

We probably don't need this `commitpending()` at all now that Python and Rust share identical store objects, but I don't want to try making that change right now.

Reviewed By: zzl0

Differential Revision: D65376829

fbshipit-source-id: c7745d0a1dbf5ad3a8313e32ec10f7d6dfc93a72