rebase: skip another remotefilelog.nodemap lookup
Summary: In remotefilelog.addrevision we have some logic to detect whether we can re-use the given file node without needing to first add it to the filelog. Previously we did a `node in self.nodemap` check, but I don't think that is necessary. We can infer the presence of the node be the fact we can look up its parents without needing to check for the node's content (thereby saving some work).

Reviewed By: quark-zju

Differential Revision: D65402132

fbshipit-source-id: d34d2742044c156cc0d3da33bd429bdf7127d966