rebase: expand --noconflict test
Summary:
Expand test-rebase-inmemory-noconflict.t to show what happens with partial rebase results after a conflict during --noconflict. If you are in single transaction mode, the entire rebase is rolled. If you are in multi-transaction mode, only the final transaction is rolled back (so you are left with a partial rebase).

I want to make sure my next diff doesn't regress this behavior.

Reviewed By: zzl0

Differential Revision: D65407232

fbshipit-source-id: 1c621a4d79178994776d5363be0fb3cf5260f6d7