manifest: remove "modified_dirs()"
Summary:
This was a specialized diff method that reported directories that gained or lost children (but excluded modified children).

D31698688 suggests the only user "debugdiffdirs" was used by watchman, but no one uses "debugdiffdirs" anymore, so let's clean it all up.

It doesn't strike me as generally useful (i.e. doesn't need to be a Manifest trait method). It seems like it could be implemented on top of diff instead of beside it.

I'm looking to make changes to manifest-tree/diff.rs, and I wanted to first remove the complexity of this unused feature.

Reviewed By: zzl0

Differential Revision: D65376837

fbshipit-source-id: c2dd455d7de4d195725ea2db3f5decc1ad894577