subtree: use dataclasses for subtree metadata
Summary:
This diff uses subtree metadata classes instead of the Python dictionaries, which will improve
readability and reduce the likelihood of errors. For example, we can use `branch.from_commit`
instead of `branch["from_commit"]`, the error in the former style can be caught by Python
type checker, but the error in the later style can only be found in runtime.

Reviewed By: muirdm

Differential Revision: D65424321

fbshipit-source-id: c6690288e631f3e05b54427c736f7c3acfbc80dd