manifest-tree: fix progress bar for "diff" operation
Summary:
Now that the diff is parallelized, the "current depth" progress bar doesn't work well since the depth jumps around depending on ordering between threads. For now switch to a simpler, albeit less useful, approach that shows the number of trees we have processed so far.

I feel like there must be a way to get an approximate average/weighted depth across the entire operation, but I don't want to spend time on this right now.

Reviewed By: zzl0

Differential Revision: D65454850

fbshipit-source-id: d14cd6eff32d71e392f5edbaba1d5f62da131101